Moved global strings into their own header
